server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
skyfield==1.48
sgp4==2.23
numpy==2.1.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0